+++ src/main/java/com/kosta/springbootproject/admincontroller/TaeController.java
@@ -7,17 +7,12 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
 
-import org.apache.poi.hssf.util.HSSFColor;
-import org.apache.poi.ss.usermodel.BorderStyle;
-import org.apache.poi.ss.usermodel.CellStyle;
-import org.apache.poi.ss.usermodel.HorizontalAlignment;
 import org.apache.poi.xssf.usermodel.XSSFWorkbook;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -407,7 +402,7 @@ public class TaeController {
 	        
 	        try {
 	        	
-	        	XSSFWorkbook workbook = companyService.listExcelDownload(param, model, pagevo);
+	        	XSSFWorkbook workbook = classesService.listExcelDownload(param, model, pagevo);
 	        	
 	            response.reset();
 	            response.setHeader("Content-Disposition", "attachment;filename=kosta_history.xls");
