+++ src/main/java/com/kosta/springbootproject/adminservice/CertificateService.java
@@ -6,7 +6,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.kosta.springbootproject.model.Certificate;
-import com.kosta.springbootproject.model.ClassRoom;
 import com.kosta.springbootproject.persistence.CertificateRepository;
 
 @Service
