+++ src/main/java/com/kosta/springbootproject/adminservice/LectureService.java
@@ -1,14 +1,11 @@
 package com.kosta.springbootproject.adminservice;
 
 import java.util.List;
-import java.util.Optional;
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
-import com.kosta.springbootproject.model.Course;
 import com.kosta.springbootproject.model.Lecture;
 import com.kosta.springbootproject.model.PageVO;
 import com.kosta.springbootproject.persistence.LectureRepository;
