+++ src/main/resources/templates/admin/classesDetail.html
@@ -81,6 +81,7 @@
 	    <input type="checkbox" name="classRecommend" th:checked="${#strings.toString(classes.classRecommend)}=='true'">
 	    </div>
 	    <button type="submit" class="btn btn-success">저장</button>
+	    <button type="button" class="btn btn-success" onclick = "location.href = '/admin/classesList'">취소</button> 
 	</form>
 	</div>
 </body>
