+++ src/main/resources/templates/admin/classeslist.html
@@ -5,12 +5,37 @@
 <head>
 <meta charset="UTF-8">
 <title>강의리스트</title>
+
+<style>
+#filterWrap{
+    width: 90%;
+    margin: 10px auto;
+}
+#filter1{
+float: left;
+width: 30%;
+box-sizing: border-box;
+}
+#filter2{
+float: left;
+width: 30%;
+margin-left: 5%;
+box-sizing: border-box;
+}
+#filter3{
+float: right;
+width: 30%;
+box-sizing: border-box;
+}
+</style>
+
 </head>
 <body>
 
 	<div class="panel panel-default" layout:fragment="content1">
+
 	
-		<h1>강의리스트</h1>
+		<h1>강의</h1>
 
 		<p>[[${resultMessage}]]</p>
 		
@@ -19,34 +44,41 @@
 		<div id="divContentline"></div>
 	<!-- 메뉴분리용 div 끝-->
 	
+<div id = "filterWrap">
+	<div id = "filter1">
+		<div class="container">
+		  <p>훈련대상:</p>
+		  <div class="btn-group" id="searchType1">
+		    <button class="btn btn-primary" value="">전체</button>
+		    <button class="btn btn-primary" value="채용예정자">채용예정자</button>
+		    <button class="btn btn-primary" value="재직자">재직자</button>  
+		  </div>
+	</div>
+	</div>
 	
-<div class="container">
-  <p>훈련대상:</p>
-  <div class="btn-group" id="searchType1">
-    <button class="btn btn-primary" value="">전체</button>
-    <button class="btn btn-primary" value="채용예정자">채용예정자</button>
-    <button class="btn btn-primary" value="재직자">재직자</button>  
-  </div>
+	<div id = "filter2">	
+		<div class="container">
+		  <p>강의상태:</p>
+		  <div class="btn-group" id="searchType1">
+		    <button class="btn btn-primary" value="">전체</button>
+		    <button class="btn btn-primary" value="가산">가산</button>
+		    <button class="btn btn-primary" value="오리">오리</button>  
+		  </div>
+		</div>
+	</div>
+	
+	<div id = "filter2">
+		<div class="container">
+		  <p>강의장:</p>
+		  <div class="btn-group" id="searchType1">
+		    <button class="btn btn-primary" value="">전체</button>
+		    <button class="btn btn-primary" value="가산">가산</button>
+		    <button class="btn btn-primary" value="오리">오리</button>  
+		  </div>
+		</div>
+	</div>
 </div>
-
-<div class="container">
-  <p>강의상태:</p>
-  <div class="btn-group" id="searchType1">
-    <button class="btn btn-primary" value="">전체</button>
-    <button class="btn btn-primary" value="가산">가산</button>
-    <button class="btn btn-primary" value="오리">오리</button>  
-  </div>
-</div>
-
-<div class="container">
-  <p>강의장:</p>
-  <div class="btn-group" id="searchType1">
-    <button class="btn btn-primary" value="">전체</button>
-    <button class="btn btn-primary" value="가산">가산</button>
-    <button class="btn btn-primary" value="오리">오리</button>  
-  </div>
-</div>
-
+	
 
 
 	<!-- 메뉴분리용 div -->
@@ -68,10 +100,17 @@
 			<button id="searchBtn">조건검색</button>
 		</div>
 
+	<!-- 엑셀 다운용 코드 -->
+		<form th:action="@{/admin/exceldownload}" method="get">
+	    <input type='hidden' name='type' th:value="${pagevo.type}">
+	    <input type='hidden' name='keyword' th:value="${pagevo.keyword}">
+	    <button type="submit" class="label label-default"><p>test</p></button>
+		</form>
+	<!-- 엑셀 다운용 코드 끝-->
 
-		<h3>
-			<a class="label label-default" th:href="@{/admin/classesInsert}">등록</a>
-		</h3>
+		<div class="pull-right">
+		<h3><a class="label label-default" th:href="@{/admin/classesInsert}">등록</a></h3>
+		</div>
 
 		<table class="table table-striped table-bordered table-hover">
 			<tr>
@@ -85,7 +124,6 @@
 			</tr>
 			<tr th:each="classes:${classeslist}">
 				<td th:text="${classes.lecture.course.subject.subName}">주제명</td>
-				<!-- <td th:text="${classes.lecture.course.courseName}">강의명</td> -->
 				<td><a th:href="@{'/admin/classesDetail/'+${classes.classNo}}" th:text="${classes.lecture.course.courseName}"></a></td>
 				<td th:text="${classes.teacher.teacherName}">강사명</td>
 				<td th:text="${classes.classOpenDate} + '~'+ ${classes.classCloseDate}">기간</td>
