+++ src/main/resources/templates/admin/companyDetail.html
@@ -34,6 +34,7 @@
 		<br>
 
 		<button type="submit" class="btn btn-success">저장</button>
+		<button type="button" class="btn btn-success" onclick = "location.href = '/admin/companyList'">취소</button> 
 	</form>
 
 </div>
