+++ src/main/resources/templates/admin/courseDetail.html
@@ -87,7 +87,7 @@
 		
 			
 		<button type="submit" class="btn btn-success">저장</button>
-		<button type="button" class="btn btn-success" onclick = "location.href = '/admin/courseDetail'">취소</button>
+		<button type="button" class="btn btn-success" onclick = "location.href = '/admin/courseList'">취소</button>
 	</form>
 
 <script>
