+++ src/main/resources/templates/admin/courseList.html
@@ -30,15 +30,9 @@
   </div>
 </div>
 
-
-
-
 	<!-- 메뉴분리용 div -->
 		<div id="divContentline"></div>
 	<!-- 메뉴분리용 div 끝-->
-	
-
-
 
 		<div id="searchType">
 			<select>
@@ -54,9 +48,10 @@
 			<button id="searchBtn">조건검색</button>
 		</div>
 		
-		<h3>
-			<a class="label label-default" th:href="@{/admin/courseInsert}">등록</a>
-		</h3>
+		
+		<div class="pull-right">
+		<h3><a class="label label-default" th:href="@{/admin/courseInsert}">등록</a></h3>
+		</div>
