+++ src/main/resources/templates/admin/lectureDetail.html
@@ -65,6 +65,7 @@
 
 			
 		<button type="submit" class="btn btn-success">저장</button>
+		<button type="button" class="btn btn-success" onclick = "location.href = '/admin/lectureList'">취소</button> 
 	</form>
 
 </div>
