+++ src/main/resources/templates/admin/lectureList.html
@@ -27,9 +27,9 @@
 		</div>
 
 
-	<h3>
-			<a class="label label-default" th:href="@{/admin/lectureInsert}">등록</a> 
-		</h3>
+	<div class="pull-right">
+	<h3><a class="label label-default" th:href="@{/admin/lectureInsert}">등록</a></h3>
+	</div>
