+++ src/main/resources/templates/admin/teacherDetail.html
@@ -33,6 +33,7 @@
 		</div>
 			
 		<button type="submit" class="btn btn-success">수정</button>
+		<button type="button" class="btn btn-success" onclick = "location.href = '/admin/teacherList'">취소</button> 
 
 	</form>
