+++ src/main/resources/templates/admin/teacherList.html
@@ -27,10 +27,9 @@
 			<button id="searchBtn">조건검색</button>
 		</div>
 		
-		
-		<h3>
-			<a class="label label-default" th:href="@{/admin/teacherInsert}">등록</a>
-		</h3>
+		<div class="pull-right">
+		<h3><a class="label label-default" th:href="@{/admin/teacherInsert}">등록</a></h3>
+		</div>
 		
 
 		<table border="1">
