+++ src/main/resources/templates/fragments/headeradmin.html
@@ -6,8 +6,8 @@ xmlns:sec="http://www.thymeleaf.org/extras/spring-security">
 <title>Insert title here</title>
 <style>
 html {
-	margin-left: 0px !important;
-	padding: 0 !important;
+  margin-left: 0px !important;
+  padding: 0 !important;
 }
 </style>
 </head>
@@ -19,23 +19,25 @@ html {
 		<div id="headerdiv">
 			<div id="headerLogoDiv">
 				<p>BOB 관리콘솔</p>
+				
 			</div>
-			<div style="float:right">
-				<p><a href="/admin/adminInfo">관리자페이지</a></p>
+				<div id="headerDivCenter">
+					<p id="time-result"></p>
+				</div>
+			<div id="headerDivRight">
+				<a href="/admin/adminInfo"><p>관리자<span class="glyphicon glyphicon-user"></span></p></a>
 			</div>
 		</div>
-
+		
+		
+<script type="text/javascript"> 
+var d = new Date(); 
+var currentDate = d.getFullYear() + "년 " + ( d.getMonth() + 1 ) + "월 " + d.getDate() + "일"; 
+/* var currentTime = d.getHours() + "시 " + d.getMinutes() + "분 " + d.getSeconds() + "초";  */
+var result = document.getElementById("time-result"); result.innerHTML = "오늘 날짜는 " + currentDate + " 입니다."; 
+</script> 
+	
 	</header>
-
-	<div th:fragment="header2">
-		<h1>header의 2번째 조각이다.</h1>
-	</div>
-
-	<div th:fragment="header3">
-		<h1>header의 3번째 조각이다.</h1>
-	</div>
-
-
 
 </body>
 </html>
